[ECOMMERCE-CONFIG] Move config processors to DependencyInjection namespace
