+++ pimcore/tests/ecommerce/DependencyInjection/Config/Processor/TenantProcessorTest.php
@@ -15,9 +15,9 @@ declare(strict_types=1);
  * @license    http://www.pimcore.org/license     GPLv3 and PEL
  */
 
-namespace Pimcore\Tests\Ecommerce\Tools\Config;
+namespace Pimcore\Tests\Ecommerce\DependencyInjection\Config\Processor;
 
-use Pimcore\Bundle\EcommerceFrameworkBundle\Tools\Config\Processor\TenantProcessor;
+use Pimcore\Bundle\EcommerceFrameworkBundle\DependencyInjection\Config\Processor\TenantProcessor;
 use Pimcore\Tests\Test\TestCase;
 use Symfony\Component\Config\Definition\Exception\InvalidConfigurationException;
